Issue #4586: Moved resources for the JavadocStyle check test
