--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/bothfiles/InputIgnored.java
@@ -1,5 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.javadoc.bothfiles;
-
-class InputIgnored
-{
-}
