--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/bothfiles/package-info.java
@@ -1,1 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.javadoc.bothfiles;
