--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/bothfiles/package.html
@@ -1,1 +1,0 @@
-Ignored
