+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocstyle/InputJavadocStyleHtmlComment.java
@@ -3,7 +3,7 @@
 // Created: 2014
 ////////////////////////////////////////////////////////////////////////////////
 
-package com.puppycrawl.tools.checkstyle.checks.javadoc;
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocstyle;
 
 /**
  * Test input for the JavadocStyleCheck.  This check is used to perform
