+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocstyle/bothfiles/package-info.java
@@ -1,0 +1,1 @@
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocstyle.bothfiles;
