+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocstyle/pkginfo/annotation/package-info.java
@@ -1,11 +1,12 @@
 /**
  * This is a valid package documentation.  <--- See the period after the
  * first sentence.
- * 
+ *
  * <p>
  * hurray for javadocs in html
  * <br>
  * with a legacy non-closed br element
  * </p>
  */
-package com.puppycrawl.tools.checkstyle.checks.javadoc.pkginfo.valid;
+@Deprecated
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocstyle.pkginfo.annotation;
