+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocstyle/pkginfo/invalidinherit/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@inheritDoc} Where are we inheriting from
+ */
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocstyle.pkginfo.invalidinherit;
