--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/pkginfo/invalidformat/package-info.java
@@ -1,4 +1,0 @@
-/**
- * blah blah
- */
-package com.puppycrawl.tools.checkstyle.checks.javadoc.pkginfo.invalidformat;
