--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/pkginfo/invalidinherit/package-info.java
@@ -1,4 +1,0 @@
-/**
- * {@inheritDoc} Where are we inheriting from
- */
-package com.puppycrawl.tools.checkstyle.checks.javadoc.pkginfo.invalidinherit;
